+++ src/log.ts
@@ -1,4 +1,3 @@
-import { Context, EventData } from './core'
 
 // side effects for log functionality
 
@@ -6,17 +5,8 @@ export const warn = (source: string, description: string) => console.warn(`sourc
 
 export const error = (source: string, description: string) => console.error(`source: ${source}, description: ${description}`)
 
-export const onDispatch = (ctx: Context, ev: EventData) => {
-  console.log(`dispatched input '${ev[1]}' from space '${ev[0]}' with parameters: '${ev[2]}' and '${ev[3]}'`)
-  if (ctx.components[ev[0]]) {
-    console.log(ctx.components[ev[0]].state)
-  } else {
-    console.log(`The space ${ev[0]} has removed`)
-  }
-}
 
 export const logFns = {
-  onDispatch,
   warn,
   error,
 }